fix: type conversion error

+++ fivekmrun_app_flutter/lib/state/strava_activity_model.dart
@@ -1,7 +1,7 @@
 import 'package:strava_flutter/domain/model/model_detailed_activity.dart';
 
 class FastestSplitSummary {
-  final int elapsedTime;
+  final double elapsedTime;
   final double distance;
 
   FastestSplitSummary({
